+++ bot.py
@@ -1,88 +1,50 @@
 #!/bin/python3
 # bot.py
-# pylint: disable=import-error
+
+# python libs
 import bot_imports
 import sys
 import os
-import random
 import traceback
+import random
 import json
 import asyncio
-import discord
 from datetime import datetime
+# additional libs
+import discord
 from dotenv import load_dotenv
 from discord.ext import commands
 from discord.utils import find
-from ext.consolemod import style
-from ext.logcfg import get_logger, logging
+# python external files
 from ext.imports_share import log, bot, get_prefix
+from ext.const import statusLs, LASTWRDFILE, STRFILE, SETFILE, slog, nlog, hint, logging, cmdHdlLogger, eventLogger, style
+from ext import excepts
 import easteregg
 from ext.chat import chat
 load_dotenv()
-print(' >> Defining constant variables...')
-statusLs = [
-    '2020 Best discord bot: Merlin', 'PyPI', 'Github', 'Repl.it', 'Minecraft', 'Windows Whistler OOBE', 'GitLab', 'readthedocs.io', 'NoCopyrightSounds', 'Discord',
-    'Recursion', 'F0rk B0mbs', 'Different 𝗞𝗶𝗻𝗱𝘀 𝘖𝘧 𝙲𝚑𝚊𝚛𝚊𝚌𝚝𝚎𝚛𝚜', 'sudo rm -rf / --no-preserve-root', 'rd/s/q %windir%', 'typing "exit" in linux init=/bin/bash',
-    'Hello, world!', 'Oracle Virtualbox VMs', 'VMware', 'Quick EMUlator (QEMU)', 'Global Information Tracker', 'Goddamn Idiotic Truckload of sh*t',
-    'Arch Linux', 'Manjaro Linux', 'Microsoft Windows 10', 'Canonical Ubuntu', 'Kubuntu and Xubuntu', 'Linux Mint', 'Pop!_OS', 'OpenSUSE', 'Elementry OS', 'MX Linux', 'Debian', 'BSD',
-    'Nothing', 'Status', 'what Merlin is playing', 'Twitter', 'StackOverflow', 'Mozilla Firefox', 'Visual Studio Code', 'zsh', 'fish', 'dash', 'mc (Midnight Commander)',
-    'Ruby On Rails', 'Python', 'JavaScript', 'Node.js', 'Angular', 'Assembly', 'C++ (see ga ga)', 'C', 'Docker', 'Java', 'ps1', 'Nim', 'Markdown', 'HTML', 'CSS', 'Perl', 'C#', 'R', 'Pascal'
-]
-cogs = lastmsg = []
+cogs = []
 for cog in os.listdir('cogs/'):
     if cog.endswith('.py'):
         cogs.append(cog[:-3])
 # token is stored inside ".env"
-TOKEN, LASTWRDFILE, SETFILE = os.getenv('DISCORD_TOKEN'), "data/lastword.json", "data/settings.json"
+TOKEN = os.getenv('DISCORD_TOKEN')
+lastword = stringTable = None
 try:
-    lastword, stringTable = json.load(open(LASTWRDFILE, 'r')), json.load(open('ext/wrds.json', 'r'))
-except FileNotFoundError:
-    lastword = []
-    open(LASTWRDFILE, 'w').close()
+    lastword    = json.load(open(LASTWRDFILE, 'r'))
+    stringTable = json.load(open(STRFILE, 'r'))
+except json.JSONDecodeError:
+    lastword = {}
+    f = open(LASTWRDFILE, 'w')
+    f.write("{}")
+    f.close()
     stringTable = json.load(open('ext/wrds.json', 'r'))
-print(' >> Configuring bot...')
-logger, eventLogger, cmdHdlLogger = get_logger('Merlin'), get_logger('EVENT'), get_logger('CMDHDL')
-logging.basicConfig(filename='discordbot.log', level=15, format='[%(asctime)s]%(levelname)s - %(name)s: %(message)s')
-HINT_LEVEL_NUM = 17
-logging.addLevelName(HINT_LEVEL_NUM, "HINT")
 
 
-def hint(self, message, *args, **kws):
-    """hint logging level"""
-    if self.isEnabledFor(HINT_LEVEL_NUM):
-        # Yes, logger takes its '*args' as 'args'.
-        self._log(HINT_LEVEL_NUM, message, args, **kws)
+<<<<<<< HEAD
 
-
+=======
+>>>>>>> nightly
 setattr(logging.Logger, 'hint', hint)
-
-
-def slog(message: str):
-    """sub log"""
-    print(' >> ' + message)
-    logger.hint(message)
-
-
-def nlog(message: str):
-    """new line long"""
-    print('\n==> ' + message)
-    logger.info(message)
-
-
-def cmd_handle_log(message: str):
-    """logging function for command handling"""
-    print('[CMDHDL]\t' + message)
-    cmdHdlLogger.info(message)
-
-
-def event_log(message: str):
-    print('[EVENT]\t' + message)
-    eventLogger.info(message)
-
-
-def cmd_handle_warn(message: str):
-    print(style.orange + message + style.reset)
-    cmdHdlLogger.warning(message)
 
 
 settings = json.load(open(SETFILE))
@@ -91,21 +53,97 @@ settings = json.load(open(SETFILE))
 bot.remove_command('help')
 MODE = os.getenv('MODE')
 
+slog("Adding bot commands...")
 
+
+# ---------
+# background tasks
+async def status():
+    await bot.wait_until_ready()
+    while True:
+        try:
+            if not MODE or MODE == 'NORMAL':
+                activity = discord.Game(name=random.choice(statusLs))
+                await bot.change_presence(status=discord.Status.online, activity=activity)
+            elif MODE == 'DEBUG':
+                activity = discord.Activity(type=discord.ActivityType(
+                    3), name="windowsboy111 debugging me")
+                await bot.change_presence(status=discord.Status.idle, activity=activity)
+            elif MODE == 'FIX':
+                await bot.change_presence(status=discord.Status.dnd, activity=discord.Activity(type=discord.ActivityType(3), name="windowsboy111 fixing me"))
+            await asyncio.sleep(30)
+        except Exception:
+            pass
+
+
+async def update():
+    global settings, stringTable, lastword
+    lastword, stringTable = json.load(
+        open(LASTWRDFILE, 'r')), json.load(open(STRFILE, 'r'))
+    settings.update(json.load(open(SETFILE, 'r')))
+    json.dump(settings, open(SETFILE, 'w'))
+
+
+async def task_update():
+    await update()
+    await asyncio.sleep(60)
+
+bot.loop.create_task(status())
+bot.loop.create_task(task_update())
+
+
+@bot.command(name='reboot', aliases=['restart'], hidden=True)
+@commands.is_owner()
+async def cmd_reboot(ctx):
+    global exitType
+    print('Bot going to log out in 10 seconds [owner disc rq] type: reboot')
+    await log('***__WARNING! BOT WILL RESTART IN 10 SECONDS!__***')
+    await ctx.send('Bot will restart in 10 seconds.')
+    await asyncio.sleep(10)
+    await ctx.send('Logging out...')
+    await log('Logging out...')
+    print('Logging out...')
+    exitType = 1
+    await bot.logout()
+
+
+@bot.command(name='shutdown', aliases=['stop', 'sdwn', 'kthxbai', 'halt'], hidden=True)
+@commands.is_owner()
+async def cmd_shutdown(ctx):
+    global exitType
+    nlog('Bot going to log out in 10 seconds [owner disc rq] type: shutdown')
+    await log('***__WARNING! BOT WILL RESTART IN 10 SECONDS!__***')
+    await ctx.send('Bot will shutdown in 10 seconds.')
+    await asyncio.sleep(10)
+    await ctx.send('Logging out...')
+    await log('Logging out...')
+    nlog('Logging out...')
+    exitType = 2
+    await bot.logout()
+
+
+@bot.command(name='update', hidden=True)
+@commands.is_owner()
+async def cmd_update(ctx):
+    await update()
+
+
+# ------
+# events
 @bot.event
 async def on_message(message: discord.Message):
-    global lastmsg
     if await easteregg.easter(message):
         return
     try:
         global lastword
         lastword[f'g{message.guild.id}'][str(message.author.id)] = message.id
-    except KeyError:
+    except Exception:
         lastword[f'g{message.guild.id}'] = {message.author.id: message.id}
-    json.dump(lastword, open(LASTWRDFILE, 'w'))
     if not isinstance(message.channel, discord.DMChannel) and message.channel.name == 'merlin-chat' and not message.author.bot:
-        await message.channel.send(chat.response(message.content))
-        return 0
+        async with message.channel.typing():
+            await message.channel.send(chat.response(message.content))
+            return 0
+    await update()
     if message.content.startswith(get_prefix(bot, message)):
         msgtoSend = f'{message.author} has issued command: '
         print(msgtoSend + style.green + message.content + style.reset)
@@ -115,27 +153,14 @@ async def on_message(message: discord.Message):
         except AttributeError:
             pass
         try:
-            result = await bot.process_commands(message)
-            if result:
-                try:
-                    if int(result) != 0:
-                        return int(result)
-                except Exception:
-                    pass
-                if isinstance(result, str) and result == 'no-rm':
-                    return 0
-            await message.delete()
+            await bot.process_commands(message)
             return 0
         except discord.ext.commands.errors.CommandNotFound:
             return 2
         except discord.errors.NotFound:
             return 2
         except Exception:
-            await message.channel.send(f'{message.author.mention}, there was an error trying to execute that command! :(')
-            print(traceback.format_exc())
-            return 1
-    if isinstance(message.channel, discord.channel.DMChannel):
-        return 0
+            print(traceback.format_exc(), file=sys.stderr)
 
 
 @bot.event
@@ -172,32 +197,17 @@ async def on_member_join(member: discord.Member):
 
 
 @bot.event
-async def on_guild_join(guild):
+async def on_guild_join(guild: discord.Guild):
+    sudoers = []
+    for name in [role.name for role in guild.owner.roles]:
+        if ['admin', 'mod', 'owner'] in name.lower():
+            sudoers.append(name)
     f = json.load(open(SETFILE, 'r'))
-    f[f'g{guild.id}'] = {"prefix": ['/']}
+    f[f'g{guild.id}'] = {"prefix": ['/'],
+                         "cmdHdl": {"cmdNotFound": 0}, "sudoers": sudoers}
     with open(SETFILE, 'w') as outfile:
         json.dump(f, outfile)
     return 0
-
-
-# background
-async def status():
-    await bot.wait_until_ready()
-    while True:
-        try:
-            if not MODE or MODE == 'NORMAL':
-                activity = discord.Game(name=random.choice(statusLs))
-                await bot.change_presence(status=discord.Status.online, activity=activity)
-            elif MODE == 'DEBUG':
-                activity = discord.Activity(type=discord.ActivityType(3), name="windowsboy111 debugging me")
-                await bot.change_presence(status=discord.Status.idle, activity=activity)
-            elif MODE == 'FIX':
-                await bot.change_presence(status=discord.Status.dnd, activity=discord.Activity(type=discord.ActivityType(3), name="windowsboy111 fixing me"))
-            await asyncio.sleep(30)
-        except Exception:
-            pass
-
-bot.loop.create_task(status())
 
 
 @bot.event
@@ -217,10 +227,8 @@ async def on_command_error(ctx, error):
             return
 
         # This prevents any cogs with an overwritten cog_command_error being handled here.
-        cog = ctx.cog
-        if cog:
-            if cog._get_overridden_method(cog.cog_command_error) is not None:
-                return
+        if ctx.cog and ctx.cog._get_overridden_method(ctx.cog.cog_command_error) is not None:
+            return
 
         # Anything in ignored will return and prevent anything happening.
         if isinstance(error, commands.errors.CommandNotFound):
@@ -230,12 +238,12 @@ async def on_command_error(ctx, error):
             except KeyError:
                 await ctx.send(":interrobang: :two: :x:\n<:err:740034702743830549> Command not found!\n<:warn:739838316374917171> something went wrong, please run `/settings`")
             return 2
-        if isinstance(error, commands.MissingRequiredArgument):
-            await ctx.invoke(bot.get_command('help'), cmdName=ctx.command.qualified_name)
-            return 4
+        if isinstance(error, commands.MissingRequiredArgument) or isinstance(error, commands.BadArgument):
+            return await ctx.invoke(bot.get_command('help'), cmdName=ctx.command.qualified_name)
+
         if isinstance(error, commands.errors.DisabledCommand):
-            await ctx.send(embed=discord.Embed(
-                title=f'{ctx.command} has been disabled. :lock:',
+            return await ctx.send(embed=discord.Embed(
+                title=f':no_entry: {ctx.command} has been disabled.',
                 description=f':x: `{ctx.message.content}`',
                 color=0xff0000
             ))
@@ -243,8 +251,9 @@ async def on_command_error(ctx, error):
 
         if isinstance(error, commands.errors.CommandInvokeError):
             await ctx.send(embed=discord.Embed(
-                title=':octagonal_sign: uh oh. An exception has occurred during the execution of the command',
-                description=stringTable['CommandInvokeError'].format(content=ctx.message.content),
+                title='<:err:740034702743830549> uh oh. An exception has occurred during the execution of the command',
+                description=stringTable['CommandInvokeError'].format(
+                    content=ctx.message.content),
                 timestamp=datetime.utcnow(),
                 color=0xff0000
             ))
@@ -263,6 +272,8 @@ async def on_command_error(ctx, error):
                 ':grey_question: Whoops. The discord special expression you have specified when issuing that command is invalid.'
                 ':mag: This error occurrs usually because of the bot fails to find the object.')
             return 4
+        if isinstance(error, excepts.NotMod):
+            return await ctx.send(str(error))
 
         # All other Errors not returned come here. And we can just print the default TraceBack.
         await log(f'Ignoring exception in command {ctx.message.content}:' + '\n\n```' + str(traceback.format_exc()) + '\n```', guild=ctx.guild)
@@ -303,6 +314,7 @@ async def _shutdown(ctx):
 
 def start(token=None, **kwargs):
     # login / start services
+    global exitType, settings, stringTable, lastword
     slog('Running / logging in...')
     token = token or os.getenv('DISCORD_TOKEN')
     while True:
@@ -313,7 +325,9 @@ def start(token=None, **kwargs):
         nlog('Logged out')
         break
     slog('Writing changes and saving data...')
-    json.dump(lastword, open(LASTWRDFILE, 'w'))
+    lastword, stringTable = json.load(
+        open(LASTWRDFILE, 'r')), json.load(open('ext/wrds.json', 'r'))
+    settings.update(json.load(open(SETFILE, 'r')))
     json.dump(settings, open(SETFILE, 'w'))
     if exitType == 2:
         print("\nExiting...")
@@ -324,6 +338,7 @@ def start(token=None, **kwargs):
     except PermissionError as e:
         print(f"OPERATION FAILED: {str(e)}")
         sys.exit(2)
+
 
 if __name__ == '__main__':
     start(TOKEN, bot=True, reconnect=True)